added styling to blu-ray page using styled components.
