+++ frontend/src/Pages/Collections/DvdPage.js
@@ -2,23 +2,22 @@ import React, { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { Row, Col } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
-import Product from '../components/Product/Product'
-import Loader from '../components/Loader'
-import Message from '../components/Message'
-import BluRayPaginate from '../components/BluRayPaginate'
-import { listProductsBluray } from '../actions/productActions'
+import Product from '../../components/Product/Product'
+import Loader from '../../components/Loader'
+import Message from '../../components/Message'
+import DvdPaginate from '../../components/DvdPaginate'
+import { listProductsDvd } from '../../actions/productActions'
 
-const BlurayPage = ({ match }) => {
+const DvdPage = ({ match }) => {
   const pageNumber = match.params.pageNumber || 1
   const dispatch = useDispatch()
 
-  const productListBluray = useSelector(state => state.productListBluray)
-  const { loading, error, products, page, pages } = productListBluray
+  const productListDvd = useSelector(state => state.productListDvd)
+  const { loading, error, products, page, pages } = productListDvd
 
   useEffect(() => {
-    dispatch(listProductsBluray(pageNumber))
+    dispatch(listProductsDvd(pageNumber))
   }, [dispatch, pageNumber])
-
   return (
     <>
       <Link to='/' className='btn btn-light'>
@@ -30,10 +29,10 @@ const BlurayPage = ({ match }) => {
         <Message variant='danger'>{error}</Message>
       ) : (
         <>
-          <h1>Blu-Rays</h1>
+          <h1>DVDs</h1>
           <Row>
             {products
-              .filter(product => product.category === 'Blu-Ray')
+              .filter(product => product.category === 'DVD')
               .map(product => (
                 <Col sm={12} md={6} lg={4} xl={3} key={product._id}>
                   <Product product={product} />
@@ -42,11 +41,9 @@ const BlurayPage = ({ match }) => {
           </Row>
         </>
       )}
-      <Row className='mt-3'>
-        <BluRayPaginate pages={pages} page={page} />
-      </Row>
+      <DvdPaginate pages={pages} page={page} />
     </>
   )
 }
 
-export default BlurayPage
+export default DvdPage
