+++ frontend/src/Pages/Collections/GamePage.js
@@ -1,23 +1,22 @@
 import React, { useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useSelector, useDispatch } from 'react-redux'
 import { Row, Col } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
+import Product from '../../components/Product/Product'
+import Loader from '../../components/Loader'
+import Message from '../../components/Message'
+import GamePaginate from '../../components/GamePaginate'
+import { listProductsGames } from '../../actions/productActions'
 
-import Product from '../components/Product/Product'
-import Loader from '../components/Loader'
-import Message from '../components/Message'
-import DvdPaginate from '../components/DvdPaginate'
-import { listProductsDvd } from '../actions/productActions'
-
-const DvdPage = ({ match }) => {
+const GamePage = ({ match }) => {
   const pageNumber = match.params.pageNumber || 1
   const dispatch = useDispatch()
 
-  const productListDvd = useSelector(state => state.productListDvd)
-  const { loading, error, products, page, pages } = productListDvd
+  const productListGames = useSelector(state => state.productListGames)
+  const { loading, error, products, page, pages } = productListGames
 
   useEffect(() => {
-    dispatch(listProductsDvd(pageNumber))
+    dispatch(listProductsGames(pageNumber))
   }, [dispatch, pageNumber])
   return (
     <>
@@ -30,10 +29,10 @@ const DvdPage = ({ match }) => {
         <Message variant='danger'>{error}</Message>
       ) : (
         <>
-          <h1>DVDs</h1>
+          <h1>Games</h1>
           <Row>
             {products
-              .filter(product => product.category === 'DVD')
+              .filter(product => product.category === 'GAME')
               .map(product => (
                 <Col sm={12} md={6} lg={4} xl={3} key={product._id}>
                   <Product product={product} />
@@ -42,9 +41,9 @@ const DvdPage = ({ match }) => {
           </Row>
         </>
       )}
-      <DvdPaginate pages={pages} page={page} />
+      <GamePaginate pages={pages} page={page} />
     </>
   )
 }
 
-export default DvdPage
+export default GamePage
